move footer release info to header
