+++ frontend-v3/src/app/footer/footer.component.html
@@ -1,5 +1,5 @@
 <div fxLayout="column" fxLayoutAlign="center center">
     <div class="footer" fxLayout="row" fxLayoutAlign="space-evenly center">    
-        release: {{ lastReleaseString }} (UTC)
+        
     </div>
 </div>
