+++ frontend-v3/src/app/footer/footer.component.ts
@@ -14,11 +14,5 @@ export class FooterComponent implements OnInit {
 
   constructor(private apiService: ApiService) { }
 
-  async ngOnInit(): Promise<void> { await this.setLastReleaseDate(); }
-
-  async setLastReleaseDate() {
-    let lastReleaseData: any = await this.apiService.getLastRelease();
-    let unsanitizedString = lastReleaseData.release_date;
-    this.lastReleaseString = unsanitizedString.split('T')[0];
-  }
+  ngOnInit() { }
 }
