+++ frontend-v3/src/app/header/header.component.html
@@ -4,6 +4,6 @@
         <mat-icon>signal_wifi_off</mat-icon>
     </div>
     <div class="header-primary" fxLayout="row" fxLayoutAlign="space-evenly center">
-        <a>getmybeats</a>
+        <a>v3 * release {{ lastReleaseString }}UTC</a>
     </div>
 </div>
