+++ frontend-v3/src/app/header/header.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 
+import { ApiService } from '../api-service';
+
 
 @Component({
   selector: 'app-header',
@@ -10,7 +12,15 @@ import { Component, OnInit, Input } from '@angular/core';
 export class HeaderComponent implements OnInit {
   @Input() lowBandwidthMode: boolean = false;
 
-  constructor() {}
+  lastReleaseString: string = '';
 
-  ngOnInit(): void {}
+  constructor(private apiService: ApiService) {}
+
+  async ngOnInit(): Promise<void> { await this.setLastReleaseDate(); }
+
+  async setLastReleaseDate() {
+    let lastReleaseData: any = await this.apiService.getLastRelease();
+    let unsanitizedString = lastReleaseData.release_date;
+    this.lastReleaseString = unsanitizedString.split('T')[0];
+  }
 }
